pyproject.toml
[project]
name = "build-with-ai-2025-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.16.1",
    "mcp>=1.9.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
